Initial commit
Pushing initial code for site operation.

This project is a spinoff from http://neverssl.com/
The goal of this project will be much more than a basic webpage to get online. For now, though, we will borrow this site's content for filler text, until the project has taken off more.